% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_geneset.R
\name{enrich_geneset}
\alias{enrich_geneset}
\title{rPEGEOS wrapper}
\usage{
enrich_geneset(gene_set)
}
\arguments{
\item{gene_set}{Query gene set of interest. Should be a Nx3 matrix, where column 1 is Entrez ID, column 2 is fold-change, and column 3 is p-value (recommend using FDR corrected p-values from a package like `limma` or `DESeq2`)}
}
\value{
A tibble with columns geneset, number_genes, cosine_similarity, and probability_random.
}
\description{
This function is a wrapper for a start to finish pathway enrichment analysis using tf-idf.
For a given gene set of interest, the algorithm will compute the cosine similarity between this vector and the gene sets in a given collection of pathways (for example, KEGG pathways.)
Additionally, the algorithm will generate a set of random gene sets that will be used to determine the probability of a given enrichment being random.
The output of the wrapper is a tidy data frame with all the enrichment scores that are better than 0.
}
