% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_results.R
\name{enrichment_results}
\alias{enrichment_results}
\title{Results data frame}
\usage{
enrichment_results(query_similarities, random_probability, number_genes,
  gene_names, pathway_names)
}
\arguments{
\item{query_similarities}{Similarity score for query vector}

\item{random_probability}{Probability of a given enrichment being random}

\item{number_genes}{Number of genes as described in \code{\link{pathway_genes}}}

\item{pathway_names}{Pathway names for pathway set, correspond to the rows of the tf-idf matrix.}
}
\value{
Returns the enrichment results for each pathway given the query set.
}
\description{
This function returns a tidy data frame with all the enrichment results.
}
