% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneset_selection.R
\name{geneset_selection}
\alias{geneset_selection}
\title{Select gene set source.}
\usage{
geneset_selection()
}
\description{
Function with no parameters, allows user to choose which gene set to use for enrichment.
}
