% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_genes.R
\name{pathway_genes}
\alias{pathway_genes}
\title{Genes in pathway}
\usage{
pathway_genes(gene_set, pathway_tfidf)
}
\arguments{
\item{gene_set}{Query gene set.  IDs of genes match colnames of tf-idf matrix.}

\item{pathway_tfidf}{tf-idf matrix computed with \code{\link{tfidf}}}
}
\value{
Returns vector with number of genes matched to any given pathway.
}
\description{
Returns the number of genes in the query set that are present in a given pathway.
}
