% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_sets.R
\docType{data}
\name{pathway_sets}
\alias{pathway_sets}
\title{Pathway sets}
\format{An object of class \code{list} of length 16.}
\usage{
pathway_sets
}
\description{
The pathway sets provided have been collated from multiple sources, including [CPDB](http://cpdb.molgen.mpg.de), [MSigDB](http://software.broadinstitute.org/gsea/msigdb/collections.jsp), and others. The following are the gene sets currently avaiable in `rPEGEOS`:
}
\details{
1: KEGG
2: PharmGKB
3: Small Molecule Pathway DB
4: WikiPathways
5: SignaLink
6: HumanCyc
7: BioCarta
8: Reactome
9: INOH
10: Pathway Interaction Database
11: NetPath
12: EHMN
13: Gene Ontology: Biological Processes
14: Gene Ontology: Cellular Component
15: Gene Ontology: Molecular Function
16: MSigDB Hallmark

More data sets can be appended to the current set to explore other biologies. The availability of these sources is dependent on their own maintenance.
}
\keyword{datasets}
