% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plotting tf-idf enrichments}
\usage{
plot_enrichment(results_df, random_thr, max_pathways)
}
\arguments{
\item{random_thr}{Threshold for probability of enrichment being random.  Defaults to 0.05.}

\item{max_pathways}{Number of top pathways to display.  Defaults to 25.}

\item{enrichment_results}{Results data frame}
}
\description{
Plots the tf-idf enrichments, ranked by similarity score.
}
