% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_probability.R
\name{random_probability}
\alias{random_probability}
\title{Random probability}
\usage{
random_probability(similarity_results, gene_set, fthr, pthr, num_sets,
  target_tfidf, tfidf_crossprod_mat)
}
\arguments{
\item{similarity_results}{Vector of cosine similarities}

\item{gene_set}{Query gene set}

\item{fthr}{Fold-change threshold for differential expression}

\item{pthr}{P-value threshold for differential expression}

\item{num_sets}{Number of random sets to generate}

\item{target_tfidf}{Pathway tf-idf matrix.}

\item{tfidf_crossprod_mat}{Cross-product matrix for pathway tf-idf.}
}
\value{
A vector of random probabilities for each pathway given the gene set size.
}
\description{
This function computes the cosine similarity between the set of randomized gene sets and the pathway tf-idf matrix. It calls the cosine_similarity function and returns the probability of a given cosine similarity being random.
}
