% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sets.R
\name{random_sets}
\alias{random_sets}
\title{Generate random sets}
\usage{
random_sets(number_sets, universe_size, gene_set_size)
}
\arguments{
\item{number_sets}{How many random sets to generate.  Defaults to 10,000}

\item{universe_size}{Size of total possible number of genes.  User defined, should match number of columns in tf-idf matrix.}

\item{gene_set_size}{How many genes to sample.  Should be the same size as the query vector.}
}
\value{
A randomized set of gene sets as a matrix RxN where R is the number of random sets and N is the number of possible genes.
}
\description{
\code{random_sets} returns a set of randomized gene sets to assess significance of enriched pathways.
}
