% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_similarity.R
\name{random_similarity}
\alias{random_similarity}
\title{Random similarities}
\usage{
random_similarity(random_set, tfidf_matrix, tfidf_crossprod_mat)
}
\arguments{
\item{random_set}{Matrix of randomized gene sets as defined on \code{\link{random_sets}}}

\item{tfidf_matrix}{Pathway tf-idf matrix.}

\item{tfidf_crossprod_mat}{Cross-product matrix for pathway tf-idf, computed with \code{\link{crossprod_matrix}}}
}
\value{
A similarity matrix that is PxR, where P is the number of pathways and R is the number of random sets.
}
\description{
\code{random_similarity} computes the cosine similarity between the set of randomized gene sets defined in \link{\code{random_sets}} and the pathway tf-idf matrix. It calls the \code{\link{cosine_similarity}} function.
}
\details{
As the output of function we get a matrix that is PxR, where P is the number of pathways and R is the number of randomized sets.  This implies that each *column* in the matrix corresponds to the cosine similarity between that random set and each one of the pathways.
}
