% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_sim.R
\name{sig_sim}
\alias{sig_sim}
\title{Significance of similarity}
\usage{
sig_sim(similarity_results, random_results)
}
\arguments{
\item{similarity_results}{Sparse tf-idf matrix (computed from \code{\link{cosine_similarity}})}

\item{random_results}{A vector to compare to the tf-idf matrix. Length of vector is equal to number of columns of tf-idf matrix}
}
\value{
Returns a vector with a probability of a given score being random
}
\description{
\code{sig_sim} returns the probability of a given similarity score being better than random.
}
