% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{tfidf}
\alias{tfidf}
\title{TF-IDF matrix}
\usage{
tfidf(data_matrix)
}
\arguments{
\item{data_matrix}{Sparse document-term matrix (generated with Matrix package)}
}
\value{
A sparse matrix with the computed tf-idf
}
\description{
Returns a tf-idf sparse matrix given a document-term matrix.
This matrix is a DxT matrix, where D is the set of documents in a corpus and T are the unique terms in that corpus.
The matrix reflects a count of occurences of a given term in a given document.
}
