% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_geneset.R
\name{enrich_geneset}
\alias{enrich_geneset}
\title{rPEGEOS wrapper}
\usage{
enrich_geneset(gene_set, pathway_names, tfidf_matrix, tfidf_crossproduct,
  num_random)
}
\arguments{
\item{gene_set}{Query gene set of interest.  IDs on gene set need to match colnames of tf-idf matrix.}

\item{pathway_names}{Pathway names for pathway set, correspond to the rows of the tf-idf matrix.  Can be a Nx1 or Nx2 object: if Nx1, use pathway names; if Nx2, column 1 is pathway name, column 2 is pathway set source (e.g., KEGG, reactome, GO)}

\item{tfidf_crossproduct}{Cross-product of the tf-idf matrix. Computed with \code{\link{crossprod_matrix}}.}

\item{num_random}{Number of random sets to be generated to calculate significance of enrichment.  Defaults to 10,000.}

\item{pathway_tfidf}{tf-idf matrix for pathway set. Column names are gene ids (EntrezIDs or gene symbols). Computed with \code{\link{rpegeos_tfidf}}}
}
\value{
A randomized set of gene sets as a matrix RxN where R is the number of random sets and N is the number of possible genes.
}
\description{
This function is a wrapper for a start to finish pathway enrichment analysis using tf-idf.
For a given gene set of interest, the algorithm will compute the cosine similarity between this vector and the gene sets in a given collection of pathways (for example, KEGG pathways.)
Additionally, the algorithm will generate a set of random gene sets that will be used to determine the probability of a given enrichment being random.
The output of the wrapper is a tidy data frame with all the enrichment scores that are better than 0.
}
