% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_geneset.R
\name{enrich_geneset}
\alias{enrich_geneset}
\title{rPEGEOS wrapper}
\usage{
enrich_geneset(gene_set)
}
\arguments{
\item{gene_set}{Query gene set of interest, provided as Entrez gene IDs.}
}
\value{
A randomized set of gene sets as a matrix RxN where R is the number of random sets and N is the number of possible genes.
}
\description{
This function is a wrapper for a start to finish pathway enrichment analysis using tf-idf.
For a given gene set of interest, the algorithm will compute the cosine similarity between this vector and the gene sets in a given collection of pathways (for example, KEGG pathways.)
Additionally, the algorithm will generate a set of random gene sets that will be used to determine the probability of a given enrichment being random.
The output of the wrapper is a tidy data frame with all the enrichment scores that are better than 0.
}
