% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_results.R
\name{enrichment_results}
\alias{enrichment_results}
\title{Results data frame}
\usage{
enrichment_results(query_similarities, random_probability, number_genes,
  gene_names, pathway_names)
}
\arguments{
\item{query_similarities}{Similarity score for query vector}

\item{random_probability}{Probability of a given enrichment being random}

\item{number_genes}{Number of genes as described in \code{\link{pathway_genes}}}

\item{gene_names}{Gene names as described in \code{\link{pathway_genes}}}

\item{pathway_names}{Pathway names for pathway set, correspond to the rows of the tf-idf matrix.  Can be a Nx1 or Nx2 object: if Nx1, use pathway names; if Nx2, column 1 is pathway name, column 2 is pathway set source (e.g., KEGG, reactome, GO)}
}
\value{
Returns the enrichment results for each pathway given the query set.
}
\description{
This function returns a tidy data frame with all the enrichment results.
}
