% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment.R
\name{plot_enrichment}
\alias{plot_enrichment}
\title{Plotting tf-idf enrichments}
\usage{
plot_enrichment(results_df)
}
\arguments{
\item{results_df}{Results data frame}

\item{max_pathways}{Number of top pathways to display.  Defaults to 20.}
}
\description{
Plots the tf-idf enrichments, ranked by similarity score.
}
